# actions_bridge/openapi.yaml
openapi: 3.1.0
info:
  title: Gurt Local Bridge
  version: "1.0.0"

# Replace with your live tunnel URL (no trailing slash)
servers:
  - url: https://<YOUR_TUNNEL_DOMAIN>

components:
  securitySchemes:
    BridgeToken:
      type: apiKey
      in: header
      name: X-Bridge-Token

security:
  - BridgeToken: []

paths:
  /health:
    get:
      security: []   # keep health open
      responses:
        "200": { description: OK }
  /run:
    post:
      operationId: post_run
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [cmd]
              properties:
                cmd:
                  type: string
                  enum: [setup, test, lint, typecheck, e2e-real]
            example: { "cmd": "test" }   # <= this makes Test send a body
      responses:
        "200": { description: OK }
  /read:
    get:
      operationId: get_read
      parameters:
        - in: query
          name: path
          required: true
          schema: { type: string }
      responses:
        "200": { description: OK }
  /write:
    post:
      operationId: post_write
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [path, content]
              properties:
                path: { type: string }
                content: { type: string }
            example: { "path": "README_LOCAL.md", "content": "hello" }
      responses:
        "200": { description: OK }
  /git:
    post:
      operationId: post_git
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [op]
              properties:
                op:
                  type: string
                  enum: [branch, commit, push]
                args:
                  type: object
                  additionalProperties: true
            example: { "op": "commit", "args": { "message": "test" } }
      responses:
        "200": { description: OK }
  /github/pr:
    post:
      operationId: post_github_pr
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [head]
              properties:
                title: { type: string }
                body:  { type: string }
                head:  { type: string }
                base:  { type: string }
            example: { "head": "feature/bridge-test", "title": "Bridge PR", "body": "Opened from Action" }
      responses:
        "200": { description: OK }